[MOV] auth_totp_mail{_enforce}: merge both modules
