+++ addons/auth_totp_mail/models/__init__.py
@@ -2,4 +2,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import auth_totp_device
+from . import auth_totp_rate_limit_log
+from . import res_config_settings
 from . import res_users
