+++ addons/auth_totp_mail/tests/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import test_auth_signup
 from . import test_notify_security_update_totp
 from . import test_totp
