--- addons/auth_totp_mail_enforce/__init__.py
@@ -1,4 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import controllers
-from . import models
