--- addons/auth_totp_mail_enforce/models/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import auth_totp_rate_limit_log
-from . import res_config_settings
-from . import res_users
