--- addons/auth_totp_mail_enforce/tests/__init__.py
@@ -1,3 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import test_auth_signup
